backend/requirements.txt
### 13. Requirements File (backend/requirements.txt)
```
Flask==2.3.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.2
werkzeug==2.3.0